+++ MaintenancePlanner-Core/src/main/java/com/maint/core/beans/MaintenancePlanDto.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class MaintenancePlanDto {
 	
-	private int id;
+	private Long id;
 	private String name;
-	private int machineId;
+	private Long machineId;
 }
