+++ MaintenancePlanner-Core/src/main/java/com/maint/core/beans/UserRoleTableDto.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class UserRoleTableDto {
 	
-	private int id;
-	private int userId;
-	private int userRoleId;
+	private Long id;
+	private Long userId;
+	private Long userRoleId;
 }
